requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
email-validator==2.1.0
Flask-Limiter==3.5.0
gunicorn==21.2.0